+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2026-01-29 03:57:19
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: b79e734989cbc7981d4c351ffcb9b875678725ad
+
+## [ForwardWidgets-2kuai] - 2026-01-29 03:57:19
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 89451ee825bbe270410af7c4bd3cbb2b789a5392
+
+## [ForwardWidgets-pack1r] - 2026-01-29 03:57:19
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: a7972fe97d208041d67481d0689490d0b05dce80
+
+
 ## [ForwardWidgets-huangxd] - 2026-01-28 03:25:42
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main